+++ src/qext_compare.c
@@ -342,7 +342,7 @@ int main(int _argc,const char **_argv){
   }
   if(xlength<test_win_size){
     fprintf(stderr,"Insufficient sample data (%lu<%lu).\n",
-     (unsigned long)xlength,test_win_size);
+     (unsigned long)xlength,(unsigned long)test_win_size);
     return EXIT_FAILURE;
   }
   if (nchannels==2 && downsample==1){
